pyproject.toml
[project]
name = "gnn2"
version = "0.1.0"
description = "Bootstrap environment for gnn2"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "matplotlib>=3.10.1",
    "numpy>=2.3.0",
    "psutil>=7.1.0",
    "pytest>=8.4.2",
    "pyyaml>=6.0.2",
    "torch>=2.12.0.dev20260317",
    "torchaudio>=2.11.0.dev20260317",
    "tqdm>=4.67.1",
    "torchvision>=0.26.0.dev20260317",
    "triton>=3.6.0",
]

[tool.uv]
package = false
prerelease = "allow"

[tool.uv.sources]
torch = { index = "pytorch-nightly-cu130" }
torchvision = { index = "pytorch-nightly-cu130" }
torchaudio = { index = "pytorch-nightly-cu130" }
triton = { index = "pytorch-nightly" }

[[tool.uv.index]]
name = "pytorch-nightly-cu130"
url = "https://download.pytorch.org/whl/nightly/cu130"
explicit = true

[[tool.uv.index]]
name = "pytorch-nightly"
url = "https://download.pytorch.org/whl/nightly"
explicit = true

[tool.pytest.ini_options]
testpaths = ["tests"]
